+++ src/core/db/mongodb.js
@@ -1,0 +1,34 @@
+const { MongoClient } = require('mongodb');
+
+const MONGODB_ACCOUNT = process.env.MONGODB_ACCOUNT
+const MONGODB_PASSWORD = process.env.MONGODB_PASSWORD
+
+
+const uri = `mongodb+srv://${MONGODB_ACCOUNT}:${MONGODB_PASSWORD}@light-cube-cluster.5wswq.mongodb.net/myFirstDatabase?retryWrites=true&w=majority`;
+const client = new MongoClient(uri, { useNewUrlParser: true, useUnifiedTopology: true });
+
+client.connect();
+
+class Mongo {
+    constructor(database) {
+        this.db = client.db(database);
+    }
+
+    async get_cur(collection) {
+        return this.db.collection(collection);
+    }
+
+    //transplanted from python
+    // def get_curs(self, collections: List[str]):
+    //     cursors: List[pymongo.cursor.CursorType] = [self.client[collection] for collection in collections]
+    //     return tuple(cursors)
+
+    // def get_all_curs(self):
+    //     cursors: List[pymongo.cursor.CursorType] = \
+    //         [self.client[collection] for collection in self.client.list_collection_names()]
+    //     return tuple(cursors)
+}
+
+module.exports = {
+    Mongo
+};